Started making a dynamic plot matrix for the interface.
