+++ tetrad-lib/src/main/java/edu/cmu/tetrad/data/Histogram.java
@@ -89,7 +89,7 @@ public class Histogram {
         if (!(low < high)) throw new IllegalArgumentException("Low must be less than high: " + low + " >= " + high);
 
         Node node = this.dataSet.getVariable(variable);
-        if (node == this.target) throw new IllegalArgumentException("Conditioning node may not be the target.");
+//        if (node == this.target) return;// throw new IllegalArgumentException("Conditioning node may not be the target.");
         if (!(node instanceof ContinuousVariable)) throw new IllegalArgumentException("Variable must be continuous.");
         if (this.continuousIntervals.containsKey(node))
             throw new IllegalArgumentException("Please remove conditioning variable first.");
@@ -123,11 +123,6 @@ public class Histogram {
         }
         this.continuousIntervals.remove(node);
         this.discreteValues.remove(node);
-    }
-
-    public void removeConditioningVariables() {
-        this.continuousIntervals = new HashMap<>();
-        this.discreteValues = new HashMap<>();
     }
 
     /**
@@ -245,17 +240,6 @@ public class Histogram {
         return this.target.getName();
     }
 
-    /**
-     * @return the number of bins for a continuous target.
-     */
-    public int getNumBins() {
-        if (this.target instanceof DiscreteVariable) {
-            return ((DiscreteVariable) this.target).getNumCategories();
-        } else {
-            return this.numBins;
-        }
-    }
-
     //======================================PRIVATE METHODS=======================================//
 
     private double[] getBreakpoints(List<Double> data, int numBins) {
@@ -356,7 +340,6 @@ public class Histogram {
     public Node getTargetNode() {
         return this.target;
     }
-
 }
